refactor(backend): split schemata/jobs.py into one file per domain
jobs.py previously contained ORM models for every domain sharing the jobs.db
sqlite file (Blueprint, GlobalGlyph, LensMetadata, ExperimentDefinition, Run,
ExperimentNext, PluginState), making it hard to navigate and edit.

Split into schemata/blueprint.py, experiment.py (includes the scheduling
subdomain's ExperimentNext), glyphs.py, lens.py, plugin.py, and run.py, each
importing Base from schemata/jobs.py. jobs.py now only owns Base, the
engine/session maker, and create_db_and_tables.

Cross-domain ForeignKeyConstraints (ExperimentDefinition/Run -> Blueprint) mean
all these ORM classes must share a single Base/MetaData -- verified this is a
hard requirement, not just style, since SQLAlchemy resolves string-based FK
targets within the referencing table's own MetaData registry.

To keep table registration correct regardless of module import order, the
entrypoint's schemata auto-discovery now imports every schemata submodule
first, and only afterwards invokes any discovered create_db_and_tables.
Previously it invoked create_db_and_tables as soon as it was found during
iteration, which happened to still work only because everything lived in one
file; splitting by domain would have silently dropped tables for
alphabetically-later modules (lens, plugin, run) on a fresh database.

No persistence or API-facing changes: same tables, same single job.db file,
same create_db_and_tables contract. Call sites updated to import model classes
from their new domain-specific module while continuing to use jobs.py's
shared sync_session_maker for the monkeypatch-based test seam.
