+++ backend/development.md
@@ -31,8 +31,10 @@
 When adding new code, make sure you place it in the right submodule:
 * routes: *all* backend routes are declared here. There is autodiscovery mechanism, do _not_ make submodules here; only declare routes in `routes/*.py` files.
   * when making changes to any code in the routes submodule, consult `routes/__init__.py` docstring!
-* schemata: *all* database schemata, ie, ORM classes, are declared here. There is autodiscovery mechanism, do _not_ make submodules here; only declare schemata in `routes/*.py` files.
-  * do not declare any functions in these files, only the ORM classes themselves, and the function related to discovery: `create_db_and_tables`
+* schemata: *all* database schemata, ie, ORM classes, are declared here. There is autodiscovery mechanism, do _not_ make submodules here; only declare schemata in flat `schemata/*.py` files, one file per domain (a subdomain, e.g. `experiment.scheduling`, may keep its table in its parent domain's schemata file instead of getting its own).
+  * do not declare any functions in these files, only the ORM classes themselves
+  * the jobs database (`job.db`) is shared by several domains' tables so that they live in a single sqlite file. `schemata/jobs.py` owns the `Base`, the engine/session maker, and the `create_db_and_tables` function related to discovery; every other jobs-domain schemata module (`blueprint.py`, `experiment.py`, `glyphs.py`, `lens.py`, `plugin.py`, `run.py`) imports `Base` from it so all their tables (including cross-table `ForeignKeyConstraint`s) register on the same `MetaData`. The entrypoint's schemata discovery imports every schemata submodule *before* invoking any discovered `create_db_and_tables`, precisely so this cross-module table registration is complete first -- see `entrypoint/app.py` for the mechanism.
+  * `schemata/user.py` is unrelated -- it owns a separate `user.db` database and its own `Base`/engine, not shared with the jobs modules above.
 * domain: the domain entities, related service functions, database helpers, domain dataclasses, et cetera. Most of the business logic lives here. Consult each domain's docstring in `__init__.py` to understand its role. When making *any* change to a code in a domain, consult the docstring to see if you need to make a change in the docstring itself.
   * within domain submodules there is often a `db.py` which contains helper functions to operate on top of ORMs from schemata. This module is always expected to handle automated version increments when mutating versioned entities, and to enforce authorization.
     * no ORM object may cross the `db.py` module boundary -- convert query results into a plain dataclass (e.g. `RunRecord`) before returning. Sessions, active result objects, and lazy-loaded attributes must never be handed to callers; materialize everything the caller needs before the session closes. The ORM objects misbehave when crossing thread boundaries -- which is bound to happen to return values of functions in `db.py` given we mix sync and async code and thread pools.
