+++ backend/src/forecastbox/domain/experiment/db.py
@@ -26,7 +26,7 @@ import forecastbox.schemata.jobs as _jobs_module
 from forecastbox.domain.blueprint.types import BlueprintId
 from forecastbox.domain.experiment.exceptions import ExperimentAccessDenied, ExperimentNotFound
 from forecastbox.domain.experiment.types import ExperimentDefinitionId
-from forecastbox.schemata.jobs import ExperimentDefinition, ExperimentType
+from forecastbox.schemata.experiment import ExperimentDefinition, ExperimentType
 from forecastbox.utility.auth import AuthContext
 from forecastbox.utility.db import dbRetry, executeAndCommit, querySingle
 from forecastbox.utility.time import current_time
