+++ backend/src/forecastbox/schemata/jobs.py
@@ -7,237 +7,32 @@
 # granted to it by virtue of its status as an intergovernmental organisation
 # nor does it submit to any jurisdiction.
 
-"""ORM models and database setup for the jobs database.
+"""Shared engine/session setup for the jobs database.
 
-Tables are versioned/immutable (Blueprint, ExperimentDefinition) or
-append-only with a mutable runtime state (Run). Soft-delete is
-supported on all main tables via `is_deleted`.
+The actual ORM models live in per-domain modules in this package (``blueprint.py``,
+``experiment.py``, ``glyphs.py``, ``lens.py``, ``plugin.py``, ``run.py``), all of which
+import ``Base`` from here so that every table is registered on the same ``MetaData``
+instance. This is required, not just cosmetic: several tables (e.g. ``ExperimentDefinition``,
+``Run``) declare ``ForeignKeyConstraint``s referencing ``blueprint`` by table name, and
+SQLAlchemy resolves those string-based references against the referencing table's own
+``MetaData`` registry -- so all of them must share this single ``Base``.
 
-Exposes ``create_db_and_tables`` so the entrypoint can discover and run it
-via automatic schemata iteration.
+Exposes ``create_db_and_tables`` so the entrypoint can discover and run it via automatic
+schemata iteration. See ``entrypoint/app.py`` for why it defers calling any discovered
+``create_db_and_tables`` until *all* schemata submodules have been imported: only then is
+it guaranteed that every ORM class in this package has registered its table on this
+module's ``Base.metadata``.
 # TODO for later: implement garbage collection
 """
 
-from typing import Literal
-
-from sqlalchemy import JSON, Boolean, CheckConstraint, Column, ForeignKeyConstraint, Integer, String, UniqueConstraint, create_engine
+from sqlalchemy import create_engine
 from sqlalchemy.orm import DeclarativeBase, sessionmaker
 
 from forecastbox.utility.config import config
-from forecastbox.utility.time import UTCDateTime
-
-BlueprintSource = Literal["plugin_template", "user_defined", "oneoff_execution"]
-ExperimentType = Literal["cron_schedule", "batch_execution", "external_trigger"]
-RunStatus = Literal["submitted", "preparing", "running", "completed", "failed", "unknown"]
 
 
 class Base(DeclarativeBase):
     pass
-
-
-class Blueprint(Base):
-    """Captures everything needed to execute a job.
-
-    Immutable once written; a new version is appended for each save.
-    The composite primary key is (blueprint_id, version). `source` distinguishes
-    plugin templates, user-defined blueprints, and one-off runs.
-    `parent_id` tracks lineage without pinning a version.
-    """
-
-    __tablename__ = "blueprint"
-
-    blueprint_id = Column(String(255), primary_key=True, nullable=False)
-    version = Column(Integer, primary_key=True, nullable=False)
-    created_by = Column(String(255), nullable=False)
-    created_at = Column(UTCDateTime, nullable=False)
-
-    # TODO later -- make sure entity validates this
-    source = Column(String(64), nullable=False)
-    # Optional lineage reference – deliberately no version to keep it discoverable
-    parent_id = Column(String(255), nullable=True)
-
-    display_name = Column(String(255), nullable=True)
-    display_description = Column(String(1024), nullable=True)
-    tags = Column(JSON, nullable=True)
-
-    # stores the full forecastbox.domain.blueprint.service.BlueprintBuilder as JSON
-    builder = Column(JSON, nullable=True)
-
-    fiabcore_major = Column(Integer, nullable=False)
-
-    is_deleted = Column(Boolean, nullable=False, default=False)
-
-
-class GlobalGlyph(Base):
-    """A user-defined glyph available for interpolation in all blueprint configurations.
-
-    The combination of (created_by, key) is unique, so each user may define their own
-    glyph for the same key name independently. Public glyphs (created by admins only)
-    carry an additional ``overriddable`` flag that controls resolution priority:
-    public-overriddable glyphs have the lowest priority and can be shadowed by a user's
-    own private glyph; public-nonoverridable glyphs have the highest priority and always win.
-
-    Invariant: ``overriddable`` must be NULL when ``public=False``, and non-NULL when
-    ``public=True``. This is enforced at both the schema and domain layers.
-    """
-
-    __tablename__ = "global_glyph"
-
-    global_glyph_id = Column(String(255), primary_key=True, nullable=False)
-    key = Column(String(255), nullable=False)
-    value = Column(String(1024), nullable=False)
-    public = Column(Boolean, nullable=False, default=False)
-    overriddable = Column(Boolean, nullable=True)
-    created_by = Column(String(255), nullable=False)
-    created_at = Column(UTCDateTime, nullable=False)
-    updated_at = Column(UTCDateTime, nullable=False)
-
-    __table_args__ = (
-        UniqueConstraint("created_by", "key", name="uq_global_glyph_created_by_key"),
-        CheckConstraint(
-            "(public = 1 AND overriddable IS NOT NULL) OR (public = 0 AND overriddable IS NULL)",
-            name="chk_global_glyph_overriddable",
-        ),
-    )
-
-
-class LensMetadata(Base):
-    """Generic, frontend-managed metadata attached to a lens type (e.g. skinnyWMS).
-
-    Each user owns their own row per (lens_id, lens_metadata_id) pair -- mirrors
-    GlobalGlyph's (created_by, key) pattern. ``public`` rows represent an
-    admin-provided default/template and are visible (read-only) to every caller
-    in addition to their own rows; they are never merged server-side -- the
-    frontend is responsible for combining them as needed.
-    """
-
-    __tablename__ = "lens_metadata"
-
-    lens_id = Column(String(255), primary_key=True, nullable=False)
-    lens_metadata_id = Column(String(255), primary_key=True, nullable=False)
-    created_by = Column(String(255), primary_key=True, nullable=False)
-    created_at = Column(UTCDateTime, nullable=False)
-    updated_at = Column(UTCDateTime, nullable=False)
-
-    metadata_content = Column(JSON, nullable=True)
-    public = Column(Boolean, nullable=False, default=False)
-
-
-class ExperimentDefinition(Base):
-    """Captures that a Blueprint should execute multiple times.
-
-    Immutable; composite primary key is (experiment_definition_id, version).
-    `experiment_type` is one of: cron_schedule | batch_execution | external_trigger.
-    `experiment_definition` is a JSON blob whose schema depends on the type.
-    """
-
-    __tablename__ = "experiment_definition"
-
-    experiment_definition_id = Column(String(255), primary_key=True, nullable=False)
-    version = Column(Integer, primary_key=True, nullable=False)
-    created_by = Column(String(255), nullable=False)
-    created_at = Column(UTCDateTime, nullable=False)
-
-    display_name = Column(String(255), nullable=True)
-    display_description = Column(String(1024), nullable=True)
-    tags = Column(JSON, nullable=True)
-
-    blueprint_id = Column(String(255), nullable=False)
-    blueprint_version = Column(Integer, nullable=False)
-
-    # TODO later -- make sure entity validates this
-    experiment_type = Column(String(64), nullable=False)
-    experiment_definition = Column(JSON, nullable=True)
-
-    is_deleted = Column(Boolean, nullable=False, default=False)
-
-    __table_args__ = (
-        ForeignKeyConstraint(
-            ["blueprint_id", "blueprint_version"],
-            ["blueprint.blueprint_id", "blueprint.version"],
-        ),
-    )
-
-
-class Run(Base):
-    """A single computation that has happened or is happening.
-
-    Mutable (status, outputs, error, cascade identifiers are written at runtime).
-    Composite primary key is (run_id, attempt_count); re-runs share the same `run_id`.
-    The optional `experiment_id` links this execution to an experiment.
-    `compiler_runtime_context` carries per-execution dynamic values (e.g.
-    cron tick time, batch element) that were used to resolve the spec.
-    """
-
-    __tablename__ = "run"
-
-    run_id = Column(String(255), primary_key=True, nullable=False)
-    attempt_count = Column(Integer, primary_key=True, nullable=False)
-    created_by = Column(String(255), nullable=False)
-    created_at = Column(UTCDateTime, nullable=False)
-    updated_at = Column(UTCDateTime, nullable=False)
-
-    blueprint_id = Column(String(255), nullable=False)
-    blueprint_version = Column(Integer, nullable=False)
-
-    experiment_id = Column(String(255), nullable=True)
-    experiment_version = Column(Integer, nullable=True)
-    compiler_runtime_context = Column(JSON, nullable=False)
-    experiment_context = Column(String(255), nullable=True)
-
-    # TODO later -- make sure entity validates this
-    status = Column(String(50), nullable=False)
-    outputs = Column(JSON, nullable=True)
-    error = Column(String(255), nullable=True)
-    progress = Column(String(255), nullable=True)
-
-    # Filled after successful cascade submission
-    cascade_job_id = Column(String(255), nullable=True)
-    cascade_proc = Column(Integer, nullable=True)
-
-    is_deleted = Column(Boolean, nullable=False, default=False)
-
-    __table_args__ = (
-        ForeignKeyConstraint(
-            ["blueprint_id", "blueprint_version"],
-            ["blueprint.blueprint_id", "blueprint.version"],
-        ),
-    )
-
-
-class ExperimentNext(Base):
-    """Mutable table tracking the next scheduled run time for an experiment.
-
-    Kept separate from the immutable ExperimentDefinition so that updating
-    the next-run time does not create a new version.
-    """
-
-    __tablename__ = "experiment_next"
-
-    experiment_next_id = Column(String(255), primary_key=True, nullable=False)
-    experiment_id = Column(String(255), nullable=False, unique=True)
-    scheduled_at = Column(UTCDateTime, nullable=False)
-    updated_at = Column(UTCDateTime, nullable=False)
-
-
-class PluginState(Base):
-    """Persisted install state for each configured plugin.
-
-    One row per plugin, keyed by the PluginCompositeId rendered as ``store:local``.
-    Written and updated during plugin install; never versioned.
-    """
-
-    __tablename__ = "plugin_state"
-
-    plugin_id = Column(String(255), primary_key=True, nullable=False)
-    plugin_version = Column(String(255), nullable=False)
-    updated_at = Column(UTCDateTime, nullable=False)
-    plugin_errors = Column(JSON, nullable=False, default=list)
-    excluded_templates = Column(JSON, nullable=False, default=list)
-    glyph_remapping = Column(JSON, nullable=False, default=dict)
-    template_errors = Column(JSON, nullable=False, default=dict)
-    asset_ingest_needed = Column(Boolean, nullable=False, default=True)
-    enabled = Column(Boolean, nullable=False, default=True)
 
 
 sync_url = f"sqlite:///{config.db.sqlite_jobdb_path}"
@@ -250,5 +45,10 @@ sync_session_maker = sessionmaker(sync_engine, expire_on_commit=False)
 
 
 def create_db_and_tables() -> None:
-    """Create the jobs database and all its tables on startup."""
+    """Create the jobs database and all its tables on startup.
+
+    Relies on every per-domain schemata module having already been imported (and thus
+    having registered its ORM classes on ``Base.metadata``) by the time this is called --
+    see the entrypoint's schemata discovery for how that is guaranteed.
+    """
     Base.metadata.create_all(sync_engine)
